--- src/hooks/useTask.ts
@@ -1,10 +1,0 @@
-import { useContext } from "react";
-import { TaskContext } from "../context/TaskContext";
-
-export function useTask() {
-  const context = useContext(TaskContext);
-  if (context === undefined) {
-    throw new Error("useTask must be used within a TaskProvider");
-  }
-  return context;
-}
