+++ src/types/index.ts
@@ -1,11 +1,4 @@
-import {
-  DragEndEvent,
-  DragOverEvent,
-  DragStartEvent,
-  SensorDescriptor,
-  SensorOptions,
-} from "@dnd-kit/core";
-import { RefObject, Dispatch, SetStateAction } from "react";
+import { RefObject } from "react";
 
 type Id = string | number;
 
@@ -19,7 +12,7 @@ type Task = {
   priority: string;
   label: string;
   points?: number;
-  createdDate: string | Date;
+  createdDate: string;
   dueDate: string | Date;
 };
 
@@ -63,89 +56,6 @@ type SidebarContextType = {
   headerRef: RefObject<HTMLDivElement>;
 };
 
-type KanbanContextType = {
-  createNewColumn: () => void;
-  updateColumn: (id: Id, title: string) => void;
-  deleteColumn: (id: Id) => void;
-  taskInColumn: (columnId: Id) => Task[];
-  createTask: (columnId: Id) => void;
-  updateTask: (task: Task) => void;
-  deleteTask: (id: Id) => void;
-  columns: Column[];
-  setColumns: React.Dispatch<React.SetStateAction<Column[]>>;
-  tasks: Task[];
-  setTasks: React.Dispatch<React.SetStateAction<Task[]>>;
-};
-
-type ColumnContextType = {
-  isEditing: boolean;
-  popoverOpenStates: {
-    [key: Id]: boolean;
-  };
-  setPopoverOpenStates: React.Dispatch<
-    React.SetStateAction<{ [key: Id]: boolean }>
-  >;
-  totalPoints: number;
-  handleClick: () => void;
-  handleBlur: () => void;
-  handleOnChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  handleKeyDown: (e: React.KeyboardEvent<HTMLInputElement>) => void;
-  column: Column;
-  tasksInColumn: Task[];
-};
-
-type TaskContextType = {
-  task: Task;
-  isPopoverOpen: boolean;
-  isEditingTitle: boolean;
-  setIsEditingTitle: Dispatch<SetStateAction<boolean>>;
-  isEditingPriority: boolean;
-  setIsEditingPriority: Dispatch<SetStateAction<boolean>>;
-  isEditingDescription: boolean;
-  setIsEditingDescription: Dispatch<SetStateAction<boolean>>;
-  isEditingAssignee: boolean;
-  setIsEditingAssignee: Dispatch<SetStateAction<boolean>>;
-  isEditingLabel: boolean;
-  setIsEditingLabel: Dispatch<SetStateAction<boolean>>;
-  isEditingDueDate: boolean;
-  setIsEditingDueDate: Dispatch<SetStateAction<boolean>>;
-  dueDateState: Date;
-  setDueDateState: Dispatch<SetStateAction<Date>>;
-  mouseIsOver: boolean;
-  setMouseIsOver: Dispatch<SetStateAction<boolean>>;
-  titleRef: React.RefObject<HTMLInputElement>;
-  descriptionRef: React.RefObject<HTMLTextAreaElement>;
-  assigneeRef: React.RefObject<HTMLInputElement>;
-  labelRef: React.RefObject<HTMLSelectElement>;
-  dueDateRef: React.RefObject<HTMLButtonElement>;
-  handleToggleIsEditing: (
-    setIsEditing: Dispatch<SetStateAction<boolean>>,
-  ) => void;
-  handleBlur: (setIsEditing: (isActive: boolean) => void) => void;
-  handleFieldChange: <T extends keyof Task>(field: T, value: Task[T]) => void;
-  handleKeydown: <T extends HTMLElement>(
-    e: React.KeyboardEvent<T>,
-    setIsEditing: (isActive: boolean) => void,
-  ) => void;
-  handleMouseEnter: () => void;
-  handleMouseLeave: () => void;
-  handleTogglePopover: (taskId: Id) => void;
-  updatePoints: (direction: "up" | "down") => void;
-  updatePriority: (newPriority: string) => void;
-  updateLabel: (newLabel: string) => void;
-};
-
-type DragAndDropContextType = {
-  activeColumn: Column | null;
-  activeTask: Task | null;
-  columnsIds: Id[];
-  tasksIds: Id[];
-  onDragStart: (event: DragStartEvent) => void;
-  onDragEnd: (event: DragEndEvent) => void;
-  onDragOver: (event: DragOverEvent) => void;
-  sensors: SensorDescriptor<SensorOptions>[];
-};
-
 export type {
   Task,
   TaskCardProps,
@@ -154,8 +64,4 @@ export type {
   Id,
   DialogDeleteProps,
   SidebarContextType,
-  KanbanContextType,
-  ColumnContextType,
-  TaskContextType,
-  DragAndDropContextType,
 };
